test: organize DSL tests into per-DSL folders (cutedsl/cutile/cute)
The cutedsl, cutile, and cute (fused-MoE) opt-in DSL backends had their tests
scattered across test/transformers/ (and a partial test/cutedsl/ folder). Move
each DSL's tests into a dedicated folder mirroring the existing test/cutedsl/
convention (drop the DSL prefix from filenames).

cutedsl (into existing test/cutedsl/):
- move test_cutedsl_fused_linear_cross_entropy_sm90 -> test_fused_linear_cross_entropy_sm90
- move test_cutedsl_rms_norm_fastpath -> test_rms_norm_fastpath
- move test_cutedsl_rope -> test_rope
- move test_swiglu_cutedsl -> test_swiglu
- delete test_cutedsl_rms_norm (duplicate subset of test/cutedsl/test_rms_norm.py;
  the canonical file is a strict superset with stricter bf16 dW tolerances)
- fix stale docstring path in test/cutedsl/test_backend.py

cutile (new test/cutile/ + __init__.py):
- move test_cutile_backend -> test_backend
- move test_cutile_fused_linear_cross_entropy_sm90 -> test_fused_linear_cross_entropy_sm90
- move test_cutile_rms_norm -> test_rms_norm
- move test_cutile_rope -> test_rope
- move test_cutile_swiglu -> test_swiglu
- split the 7 env-gated test_cutile_* tests out of the shared
  test/transformers/test_fused_linear_scaled_cross_entropy.py into
  test/cutile/test_fused_linear_scaled_cross_entropy.py (frontend tests stay)

cute (new test/cute/ + __init__.py):
- move test_cute_moe_autograd -> test_moe_autograd
- move test_moe -> test_moe
(test_fused_moe.py is the general Triton MoE test and stays in test/transformers/)

Makefile:
- test/test-verbose now ignore test/cutile and test/cute (and test-verbose also
  ignores test/cutedsl for consistency)
- add test-cutile (LIGER_KERNEL_IMPL=cutile) and test-cute (no env var, to keep
  the always-on registration/import-guard tests working without the lck wheel)

Import blocks in the moved rope/swiglu files were re-sorted by ruff because
test.utils is now first-party inside the packaged test folders.
