+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: test test-verbose checkstyle test-convergence test-cutedsl all serve build clean
+.PHONY: test test-verbose checkstyle test-convergence test-cutedsl test-cutile test-cute all serve build clean
 
 
 all: checkstyle test test-convergence
@@ -8,12 +8,27 @@ test:
 	python -m pytest --disable-warnings \
 		--ignore=test/convergence \
 		--ignore=test/cutedsl \
+		--ignore=test/cutile \
+		--ignore=test/cute \
 		test/
 
 # Command to run pytest for CuTe DSL kernel tests
 # Requires nvidia-cutlass-dsl: pip install -e '.[cutedsl]'
 test-cutedsl:
 	LIGER_KERNEL_IMPL=cutedsl python -m pytest --disable-warnings test/cutedsl/
+
+# Command to run pytest for cuTile kernel tests
+# Requires cuda-tile: pip install -e '.[cutile]' (or '.[cutile-tileiras]')
+test-cutile:
+	LIGER_KERNEL_IMPL=cutile python -m pytest --disable-warnings test/cutile/
+
+# Command to run pytest for the cute fused-MoE tests
+# The pure-Python registration/import-guard tests always run; the functional
+# tests need the separate liger_cute_kernels (lck) wheel and >=2 CUDA devices.
+# LIGER_KERNEL_IMPL is intentionally NOT set to 'cute': selecting it would force
+# `import liger_kernel.ops` to load the lck wheel and break the always-on tests.
+test-cute:
+	python -m pytest --disable-warnings test/cute/
 
 # Command to run pytest for correctness tests with coverage reporting and full test-duration breakdown
 test-verbose:
@@ -24,6 +39,9 @@ test-verbose:
 		--cov-config=pyproject.toml \
 		--durations=0 \
 		--ignore=test/convergence \
+		--ignore=test/cutedsl \
+		--ignore=test/cutile \
+		--ignore=test/cute \
 		test/
 
 # Command to run coverage report
