+++ test/cutile/test_rope.py
@@ -1,9 +1,18 @@
-"""Correctness tests for the CuteDSL RoPE kernel.
+"""Correctness tests for the cuTile RoPE kernel.
 
-These tests require the optional ``nvidia-cutlass-dsl`` package and an NVIDIA
-GPU; they are skipped otherwise. Numerics are checked against the existing
-Triton kernel (``LigerRopeFunction``) and HuggingFace's ``apply_rotary_pos_emb``
-using *real* rotary embeddings (whose two head-dim halves are duplicated).
+These tests require the optional ``cuda-tile`` package and an NVIDIA GPU; they
+are skipped otherwise. Numerics are checked against the existing Triton kernel
+(``LigerRopeFunction`` from ``liger_kernel.ops.rope``) and HuggingFace's
+``apply_rotary_pos_emb`` using *real* rotary embeddings (whose two head-dim
+halves are duplicated).
+
+The cuTile RoPE op has two internal code paths and the shape sweep below
+exercises both:
+
+* the **ALIGNED** 4-D ``ct.load``/``ct.store`` fast path, taken when
+  ``head_dim // 2``, ``n_q_heads`` and ``n_kv_heads`` are all powers of two; and
+* the **general** ``ct.gather``/``ct.scatter`` path, taken for any non-power-of-2
+  shape (e.g. ``head_dim=92`` -> ``head_dim_half=46``, ``n_q_heads=73``).
 """
 
 import importlib.util
@@ -11,7 +20,6 @@ import importlib.util
 import pytest
 import torch
 
-from test.utils import supports_bfloat16
 from transformers.models.llama.configuration_llama import LlamaConfig
 from transformers.models.llama.modeling_llama import LlamaRotaryEmbedding
 from transformers.models.llama.modeling_llama import apply_rotary_pos_emb
@@ -19,32 +27,34 @@ from transformers.models.llama.modeling_llama import apply_rotary_pos_emb
 from liger_kernel.ops.rope import LigerRopeFunction
 from liger_kernel.utils import infer_device
 from liger_kernel.utils import transformers_version_dispatch
+from test.utils import supports_bfloat16
 
 device = infer_device()
 
-cutedsl_available = importlib.util.find_spec("cutlass") is not None and torch.cuda.is_available()
+cutile_available = importlib.util.find_spec("cuda.tile") is not None and torch.cuda.is_available()
 
 pytestmark = pytest.mark.skipif(
-    not cutedsl_available,
-    reason="nvidia-cutlass-dsl + CUDA GPU required for CuteDSL RoPE",
+    not cutile_available,
+    reason="cuda-tile + CUDA GPU required for cuTile RoPE",
 )
 
-if cutedsl_available:
-    from liger_kernel.ops.cutedsl.ops.rope import LigerRopeCuteDSLFunction
+if cutile_available:
+    from liger_kernel.ops.cutile.ops.rope import LigerRopeFunction as LigerRopeCuTileFunction
 
 
 @pytest.mark.parametrize(
     "bsz, seq_len, num_q_heads, num_kv_heads, head_dim",
     [
+        # ---- ALIGNED path (all of head_dim//2, n_q_heads, n_kv_heads are pow2) ----
         (1, 128, 32, 32, 64),
         (2, 128, 32, 32, 64),
         # different q / k heads
         (1, 128, 32, 8, 64),
         (2, 128, 32, 8, 64),
-        # weird, non-tile-aligned shapes (exercise predicated kernel)
+        # ---- general gather/scatter path (non-pow2 heads / head_dim//2) ----
         (3, 423, 73, 213, 92),
         (3, 423, 73, 155, 92),
-        # long sequence (the perf-relevant shape)
+        # long sequence (the perf-relevant shape), ALIGNED
         (1, 8192, 32, 8, 128),
     ],
 )
@@ -62,7 +72,7 @@ if cutedsl_available:
     ],
 )
 @pytest.mark.parametrize("expand_position_ids", [True, False])
-def test_cutedsl_matches_triton_and_hf(
+def test_cutile_matches_triton_and_hf(
     bsz, seq_len, num_q_heads, num_kv_heads, head_dim, dtype, atol, rtol, expand_position_ids
 ):
     rotary_emb = transformers_version_dispatch(
@@ -80,8 +90,8 @@ def test_cutedsl_matches_triton_and_hf(
     k_hf = _k.clone().requires_grad_(True)
     q_tt = _q.clone().requires_grad_(True)
     k_tt = _k.clone().requires_grad_(True)
-    q_cu = _q.clone().requires_grad_(True)
-    k_cu = _k.clone().requires_grad_(True)
+    q_ct = _q.clone().requires_grad_(True)
+    k_ct = _k.clone().requires_grad_(True)
 
     pos_ids = torch.arange(seq_len, device=device, dtype=torch.long).unsqueeze(0)
     if expand_position_ids:
@@ -91,30 +101,31 @@ def test_cutedsl_matches_triton_and_hf(
     # forward
     hf_q, hf_k = apply_rotary_pos_emb(q_hf, k_hf, cos, sin)
     tt_q, tt_k = LigerRopeFunction.apply(q_tt, k_tt, cos, sin)
-    cu_q, cu_k = LigerRopeCuteDSLFunction.apply(q_cu, k_cu, cos, sin)
+    ct_q, ct_k = LigerRopeCuTileFunction.apply(q_ct, k_ct, cos, sin)
 
-    assert torch.allclose(cu_q, hf_q, atol=atol, rtol=rtol)
-    assert torch.allclose(cu_k, hf_k, atol=atol, rtol=rtol)
-    assert torch.allclose(cu_q, tt_q, atol=atol, rtol=rtol)
-    assert torch.allclose(cu_k, tt_k, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_q, hf_q, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_k, hf_k, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_q, tt_q, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_k, tt_k, atol=atol, rtol=rtol)
 
     # backward
     dq = torch.randn_like(hf_q)
     dk = torch.randn_like(hf_k)
     q_hf_grad, k_hf_grad = torch.autograd.grad((hf_q, hf_k), (q_hf, k_hf), (dq, dk), allow_unused=True)
-    q_cu_grad, k_cu_grad = torch.autograd.grad((cu_q, cu_k), (q_cu, k_cu), (dq.clone(), dk.clone()), allow_unused=True)
+    q_ct_grad, k_ct_grad = torch.autograd.grad((ct_q, ct_k), (q_ct, k_ct), (dq.clone(), dk.clone()), allow_unused=True)
 
-    assert torch.allclose(q_cu_grad, q_hf_grad, atol=atol, rtol=rtol)
-    assert torch.allclose(k_cu_grad, k_hf_grad, atol=atol, rtol=rtol)
+    assert torch.allclose(q_ct_grad, q_hf_grad, atol=atol, rtol=rtol)
+    assert torch.allclose(k_ct_grad, k_hf_grad, atol=atol, rtol=rtol)
 
 
 @pytest.mark.parametrize(
     "bsz, seq_len, num_q_heads, num_kv_heads, head_dim",
     [
+        # ALIGNED path
         (1, 128, 32, 32, 64),
         (2, 128, 32, 8, 64),
         (1, 2048, 32, 8, 128),
-        # token-fallback head_dim (no TMA): fast path must still be correct
+        # general gather/scatter path
         (3, 423, 73, 155, 92),
     ],
 )
@@ -131,17 +142,20 @@ def test_cutedsl_matches_triton_and_hf(
         (torch.float16, 1e-2, 1e-5),
     ],
 )
-def test_cutedsl_backward_contiguous_grad(bsz, seq_len, num_q_heads, num_kv_heads, head_dim, dtype, atol, rtol):
-    """Exercise the no-copy backward fast path.
+def test_cutile_backward_contiguous_grad(bsz, seq_len, num_q_heads, num_kv_heads, head_dim, dtype, atol, rtol):
+    """Exercise the functional ``rope_forward`` + ``rope_backward`` pair directly
+    with an already-contiguous incoming gradient.
 
-    When the incoming gradient is already ``(bsz, n_head, seq, hd)``-contiguous
-    (e.g. eager attention, or any model that calls ``.contiguous()`` after RoPE),
-    ``rope_backward`` rotates it in place with the TMA kernel in ``SEQ_INNER`` mode
-    instead of materialising a token-major copy. The autograd-driven test above
-    only feeds transpose-view grads (the fallback path), so this covers the fast
-    path directly and checks numerics against HuggingFace.
+    The autograd-driven test above only feeds transpose-view grads. Here we drive
+    the functional API by hand -- ``rope_forward`` returns the exact tile/config
+    parameters (``ALIGNED``, ``TILE_QH``, ``TILE_KH``, ``TILE_HD`` ...) that
+    ``rope_backward`` needs, so this mirrors real autograd usage but lets us pass a
+    genuinely ``(bsz, n_head, seq, hd)``-contiguous gradient (e.g. eager attention,
+    or any model that calls ``.contiguous()`` after RoPE). Checks numerics against
+    HuggingFace for both the ALIGNED and the general code paths.
     """
-    from liger_kernel.ops.cutedsl.ops.rope import rope_backward
+    from liger_kernel.ops.cutile.ops.rope import rope_backward
+    from liger_kernel.ops.cutile.ops.rope import rope_forward
 
     rotary_emb = transformers_version_dispatch(
         "4.48.0",
@@ -161,25 +175,47 @@ def test_cutedsl_backward_contiguous_grad(bsz, seq_len, num_q_heads, num_kv_head
 
     hf_q, hf_k = apply_rotary_pos_emb(q_hf, k_hf, cos, sin)
 
-    # A genuinely (bsz, n_head, seq, hd)-CONTIGUOUS gradient triggers the fast path.
+    # Run the functional forward to obtain the exact tile/config params backward needs.
+    _, _, saved_cos, saved_sin, cos_bs, aligned, tile_qh, tile_kh, tile_hd, original_dtype = rope_forward(
+        _q.clone(), _k.clone(), cos, sin
+    )
+
+    # A genuinely (bsz, n_head, seq, hd)-CONTIGUOUS gradient.
     dq = torch.randn(bsz, num_q_heads, seq_len, head_dim, device=device, dtype=dtype)
     dk = torch.randn(bsz, num_kv_heads, seq_len, head_dim, device=device, dtype=dtype)
     assert dq.is_contiguous() and dk.is_contiguous()
 
     q_hf_grad, k_hf_grad = torch.autograd.grad((hf_q, hf_k), (q_hf, k_hf), (dq, dk), allow_unused=True)
-    q_cu_grad, k_cu_grad = rope_backward(dq.clone(), dk.clone(), cos, sin)
 
-    assert torch.allclose(q_cu_grad, q_hf_grad, atol=atol, rtol=rtol)
-    assert torch.allclose(k_cu_grad, k_hf_grad, atol=atol, rtol=rtol)
+    q_ct_grad, k_ct_grad = rope_backward(
+        dq.clone(),
+        dk.clone(),
+        saved_cos,
+        saved_sin,
+        cos_bs,
+        aligned,
+        tile_qh,
+        tile_kh,
+        tile_hd,
+        original_dtype,
+        bsz,
+        seq_len,
+        num_q_heads,
+        num_kv_heads,
+        head_dim,
+    )
+
+    assert torch.allclose(q_ct_grad, q_hf_grad, atol=atol, rtol=rtol)
+    assert torch.allclose(k_ct_grad, k_hf_grad, atol=atol, rtol=rtol)
 
 
 @pytest.mark.parametrize(
     "seq_len, num_heads, head_dim",
     [
-        # TMA fast path (even head_dim)
+        # ALIGNED path (head_dim//2, num_heads all pow2)
         (128, 32, 64),
         (256, 16, 128),
-        # token fallback (head_dim=92 is not TMA-vectorizable)
+        # general gather/scatter path (non-pow2 head_dim//2 and num_heads)
         (423, 73, 92),
     ],
 )
@@ -196,15 +232,15 @@ def test_cutedsl_backward_contiguous_grad(bsz, seq_len, num_q_heads, num_kv_head
         (torch.float16, 1e-2, 1e-5),
     ],
 )
-def test_cutedsl_vision_2d_cos_sin(seq_len, num_heads, head_dim, dtype, atol, rtol):
+def test_cutile_vision_2d_cos_sin(seq_len, num_heads, head_dim, dtype, atol, rtol):
     """Vision RoPE feeds a 2-D ``(seq_len, head_dim)`` cos/sin table (no batch dim).
 
     ``liger_rotary_pos_emb_vision`` transposes q/k to 4-D but forwards cos/sin
-    *unchanged* as rank-2 tensors, so the kernel must accept a rank-2 table (the
-    CuteDSL ``rope_forward``/``rope_backward`` handle this via a ``cos.dim()==2``
-    unsqueeze). Numerics are checked against HuggingFace (fed the equivalent 3-D
-    table) and the Triton kernel (fed the same 2-D table), for both the TMA and
-    the token-fallback code paths.
+    *unchanged* as rank-2 tensors. This is a regression test: the cuTile kernel
+    previously raised ``TileTypeError: Index size 3 does not match the array rank
+    2`` on such a table because it indexes cos/sin with a 3-tuple. Numerics are
+    checked against HuggingFace (fed the equivalent 3-D table) and the Triton
+    kernel (fed the same 2-D table), for both the ALIGNED and general code paths.
     """
     rotary_emb = transformers_version_dispatch(
         "4.48.0",
@@ -214,6 +250,8 @@ def test_cutedsl_vision_2d_cos_sin(seq_len, num_heads, head_dim, dtype, atol, rt
         after_kwargs={"config": LlamaConfig(num_kv_heads=num_heads, head_dim=head_dim), "device": device},
     )
 
+    # Vision layout: q, k arrive as (seq_len, num_heads, head_dim); the wrapper
+    # unsqueezes a batch dim and transposes to (1, num_heads, seq_len, head_dim).
     _q = torch.randn((seq_len, num_heads, head_dim), device=device).to(dtype)
     _k = torch.randn((seq_len, num_heads, head_dim), device=device).to(dtype)
 
@@ -224,26 +262,28 @@ def test_cutedsl_vision_2d_cos_sin(seq_len, num_heads, head_dim, dtype, atol, rt
     k_hf = to_4d(_k).clone().requires_grad_(True)
     q_tt = to_4d(_q).clone().requires_grad_(True)
     k_tt = to_4d(_k).clone().requires_grad_(True)
-    q_cu = to_4d(_q).clone().requires_grad_(True)
-    k_cu = to_4d(_k).clone().requires_grad_(True)
+    q_ct = to_4d(_q).clone().requires_grad_(True)
+    k_ct = to_4d(_k).clone().requires_grad_(True)
 
     pos_ids = torch.arange(seq_len, device=device, dtype=torch.long).unsqueeze(0)
     cos_3d, sin_3d = rotary_emb(k_hf, pos_ids)  # (1, seq_len, head_dim)
     cos_2d, sin_2d = cos_3d[0], sin_3d[0]  # (seq_len, head_dim) -- the vision table
 
+    # forward: HF gets the 3-D table; the kernels get the bare 2-D table.
     hf_q, hf_k = apply_rotary_pos_emb(q_hf, k_hf, cos_3d, sin_3d)
     tt_q, tt_k = LigerRopeFunction.apply(q_tt, k_tt, cos_2d, sin_2d)
-    cu_q, cu_k = LigerRopeCuteDSLFunction.apply(q_cu, k_cu, cos_2d, sin_2d)
+    ct_q, ct_k = LigerRopeCuTileFunction.apply(q_ct, k_ct, cos_2d, sin_2d)
 
-    assert torch.allclose(cu_q, hf_q, atol=atol, rtol=rtol)
-    assert torch.allclose(cu_k, hf_k, atol=atol, rtol=rtol)
-    assert torch.allclose(cu_q, tt_q, atol=atol, rtol=rtol)
-    assert torch.allclose(cu_k, tt_k, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_q, hf_q, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_k, hf_k, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_q, tt_q, atol=atol, rtol=rtol)
+    assert torch.allclose(ct_k, tt_k, atol=atol, rtol=rtol)
 
+    # backward
     dq = torch.randn_like(hf_q)
     dk = torch.randn_like(hf_k)
     q_hf_grad, k_hf_grad = torch.autograd.grad((hf_q, hf_k), (q_hf, k_hf), (dq, dk), allow_unused=True)
-    q_cu_grad, k_cu_grad = torch.autograd.grad((cu_q, cu_k), (q_cu, k_cu), (dq.clone(), dk.clone()), allow_unused=True)
+    q_ct_grad, k_ct_grad = torch.autograd.grad((ct_q, ct_k), (q_ct, k_ct), (dq.clone(), dk.clone()), allow_unused=True)
 
-    assert torch.allclose(q_cu_grad, q_hf_grad, atol=atol, rtol=rtol)
-    assert torch.allclose(k_cu_grad, k_hf_grad, atol=atol, rtol=rtol)
+    assert torch.allclose(q_ct_grad, q_hf_grad, atol=atol, rtol=rtol)
+    assert torch.allclose(k_ct_grad, k_hf_grad, atol=atol, rtol=rtol)
